+++ jkplus/game/jkplus_main.c
@@ -37,6 +37,7 @@ vmCvar_t	jkcvar_serverClosedBroadcast;
 vmCvar_t	jkcvar_allowBlackNames;
 vmCvar_t	jkcvar_allowMultiDuel;
 vmCvar_t	jkcvar_allowDuelChat;
+vmCvar_t	jkcvar_noDuplicatedNames;
 
 vmCvar_t	jkcvar_voteCustomMap;
 vmCvar_t	jkcvar_voteGameplay;
@@ -68,6 +69,7 @@ static cvarTable_t	JKPlusCvarTable[] = {
 	{ &jkcvar_allowBlackNames,			"jk_allowBlackNames",		"0",					CVAR_ARCHIVE,		0, qtrue },
 	{ &jkcvar_allowMultiDuel,			"jk_allowMultiDuel",		"0",					CVAR_ARCHIVE,		0, qtrue },
 	{ &jkcvar_allowDuelChat,			"jk_allowDuelChat",			"0",					CVAR_ARCHIVE,		0, qtrue },
+	{ &jkcvar_noDuplicatedNames,		"jk_noDuplicatedNames",		"0",					CVAR_ARCHIVE,		0, qtrue },
 
 	{ &jkcvar_voteCustomMap,			"jk_voteCustomMap",			"0",					CVAR_ARCHIVE,		0, qtrue },
 	{ &jkcvar_voteGameplay,				"jk_voteGameplay",			"0",					CVAR_ARCHIVE,		0, qtrue },